// src/luogu-api.d.ts
export interface ProblemSummary {
  pid: string;
  title: string;
  difficulty: number;
  fullScore: number;
  type: string;
}

export interface Problem extends ProblemSummary {
  tags: number[];
  wantsTranslation: boolean;
  totalSubmit: number;
  totalAccepted: number;
  flag: number;
}

export interface ProblemDetails extends Problem {
  background: string;
  description: string;
  inputFormat: string;
  outputFormat: string;
  samples: [string, string][];
  hint: string;
  provider: UserSummary | TeamSummary;
  attachments: {
    size: number;
    uploadTime: number;
    downloadLink: string;
    id: string;
    fileName: string;
  }[];
  canEdit: boolean;
  limits: {
    time: number[];
    memory: number[];
  };
  showScore: boolean;
  score: number | null;
  stdCode: string;
  vjudge?: {
    origin: string;
    link: string;
    id: string;
  };
  translation?: string;
}
// ...
export interface UserSummary {
  uid: number;
  name: string;
  slogan: string | null;
  badge: string | null;
  isAdmin: boolean;
  isBanned: boolean;
  color: string;
  ccfLevel: number;
  background: string;
  isRoot?: true;
}
// ...
export interface TeamSummary {
  id: number;
  name: string;
  isPremium: boolean;
}